:sparkles: Adicionando validações nos crud's

+++ controller/livrosController.ts
@@ -1,8 +1,10 @@
 import Connection from "../connection/connection";
+import { Autor } from "../entities/autor";
 import { Livro } from "../entities/livro";
 
 export class LivrosController {
   livrosRepository = Connection.getRepository(Livro);
+  autoresRepository = Connection.getRepository(Autor);
 
   listar = async (req: any, res: any) => {
     try {
@@ -19,44 +21,76 @@ export class LivrosController {
       const livro = await this.livrosRepository.findOneBy({
         id_livro: id_livro,
       });
-      return res.status(200).json(livro);
+      if (livro){
+        return res.status(200).json(livro);
+      }else{
+        return res.status(200).json({message: "Livro não encontrado!"})
+      }
     } catch (error) {
       return res.status(500).json({ error });
     }
   };
 
   inserir = async (req: any, res: any) => {
-    try {
-      const livro = await this.livrosRepository.create(req.body);
-      const results = await this.livrosRepository.save(livro);
-      return res.status(200).json({
-        message: "Livro Inserido com Sucesso",
-        payload: results,
-      });
-    } catch (error) {
-      return res.status(500).json({ error });
+    if(req.body.nome_livro && req.body.editora && req.body.ano_publicacao && req.body.id_autor){
+      try {
+        const livro = await this.livrosRepository.create(req.body);
+        const results = await this.livrosRepository.save(livro);
+        return res.status(200).json({
+          message: "Livro Inserido com Sucesso",
+          payload: results,
+        });
+      } catch (error) {
+        return res.status(400).json({ message: "Autor Não Encontrado" });       
+      }
+    }else{
+      return res.status(400).json({ message: "Preencha os campos corretamente" })
     }
   };
 
   deletar = async (req: any, res: any) => {
-    try {
-      const livro = await this.livrosRepository.delete(req.params.id_livro);
-      return res.status(200).json({ message: "Livro Excluído com Sucesso" });
-    } catch (error) {
-      return res.status(500).json({ error });
+    const livros = await this.livrosRepository.findOneBy({
+      id_livro: req.params.id_livro,
+    });
+    if(!livros){
+      return res.status(404).json({message: "Livro não encontrado"})
+    }else {
+      try {
+        const livro = await this.livrosRepository.delete(req.params.id_livro);
+        return res.status(200).json({ message: "Livro Excluído com Sucesso" });
+      } catch (error) {
+        return res.status(500).json({ error });
+      }
     }
   };
 
   atualizar = async (req: any, res: any) => {
+    const { id_livro } = req.params;
     try {
-      const cliente = await this.livrosRepository.findOne(req.params.id_livro);
-      this.livrosRepository.merge(cliente, req.body);
-      const result = await this.livrosRepository.save(cliente);
-      return res
-        .status(200)
-        .json({ message: "Livro atualizado com sucesso", payload: result });
+      const cliente_update = await this.livrosRepository.update(
+        id_livro,
+        req.body
+      );
+      const livros = await this.livrosRepository.findOneBy({
+        id_livro: req.params.id_livro
+      })
+      if(livros){
+        if(req.body.nome_livro || req.body.editora || req.body.ano_publicacao || req.body.id_autor){
+          return res.status(200).json({
+            message: "Livro atualizado com sucesso."
+          });
+        }else{
+          return res.status(400).json({message: "Preencha os campos corretamente"})
+        }
+      }else{
+        return res.status(200).json({message: "Livro não encontrado"})
+      }
+      
     } catch (error) {
-      return res.status(500).json({ error });
+      return res.status(500).json({
+        message: "Erro ao tentar atualizar livro.",
+        info: error,
+      });
     }
   };
 }
